kotizo-backend/core/permissions.py
from rest_framework.permissions import BasePermission

class IsAdminKotizo(BasePermission):
    def has_permission(self, request, view):
        return bool(request.user and request.user.is_authenticated and request.user.is_staff)

class IsVerifie(BasePermission):
    def has_permission(self, request, view):
        return bool(request.user and request.user.is_authenticated and request.user.identite_verifiee)

class IsBusiness(BasePermission):
    def has_permission(self, request, view):
        return bool(request.user and request.user.is_authenticated and request.user.niveau == 'business')