kotizo-backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
django-environ==0.11.2
psycopg2-binary==2.9.10
dj-database-url==2.3.0
redis==5.2.1
django-redis==5.4.0
celery==5.4.0
celery[redis]==5.4.0
django-celery-beat==2.7.0
flower==2.0.1
requests==2.32.3
google-generativeai==0.8.3
firebase-admin==6.6.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-ratelimit==4.1.0
argon2-cffi==23.1.0
sentry-sdk==2.19.2
gunicorn==23.0.0
whitenoise==6.8.2
pillow==11.1.0
python-decouple==3.8
phonenumbers==8.13.50